pyInterface: Add bindings for nBodyPhaseSpaceGen.
Make the nBodyPhaseSpace class available in the python module. Note that this
has not been thoroughly tested, but at least generating some events seems to
yield reasonable results. Note that the random seed for the generator cannot be
changed at the moment, as the randomNumberGenerator class is not exposed to
python.