amplitude: Add possibility for consistent subDecays.
Introduce a new function in the isobarDecayTopology, subDecayConsistent, which
returns a sub decay which is passing the consistency checks and can be used to
instantiate and calculate amplitudes. To do this, also add a new vertex class,
the nonInteractionVertex, where the incoming particle does not interact and is
also the outgoing particle. It is then used as a production vertex to make the
sub decay consistent.