amplitude: Make vertex aware of decay.
Introduce a weak_ptr to a decay topology as member variable in the
interactionVertex. It has to be a weak_ptr there will be a reference cycle
otherwize (decayTopology -> vertexPtr -> decayTopologyPtr). The weak_ptr is set
when an amplitude is constructed from a decayTopology. The whole thing is
fickle: as soon vertices get passed between amplitudes or amplitudes are
constructed of subDecays, the weak_ptr gets overwritten, but for now it serves
its purpose and the general scheme can be adapted if needed.