amplitude: Revise phaseSpaceIntegral class.
Completely revise the phaseSpaceIntegral class, which should now be somewhat
functional (albeit not tested) for arbitrary topologies. Move the numerical
integration to a Monte Carlo based one. The integrand is obtained from the
original decayTopology so it should work in a general manner. There are a
number of parameters, which are currently set with const static variables,
which should be moved to a configuration file. There is still one outstanding
problem: the class will produce separate integral tables for different spin
projections M with the same integral values in (save statistical fluctuations),
this should be fixed.