amplitude: Revice phaseSpaceIntegral for performance.
Rewrite the phaseSpaceIntegral class almost completely to bring performance to
a reasonable level. In the last incarnation, the integral files were opened and
closed for every event, which was nonsense of course. Now, the
phaseSpaceIntegral class is actually just a bookkeeper which manages
integralTableContainer objects. These contain most of the code which was in the
phaseSpaceIntegral before plus some new nifty features. This means, the
integral tables themselves are held in memory for the duration of the
calculations, without any need to reload them from disk. This improves the
performance by orders of magnitude up to a point where one could consider it
usable.

A big chunk which is still missing is the external setting of parameters (i.e.
via a config file). This is a little bit tricky because it involves changing
all the executables which (might) use the phaseSpaceIntegral, but it will be
done in the future.