Delete gui.
The scripts in src/rootscripts were compiled in CMake, which should not be the
case. However, the function in the scripts are used by the gui package for some
reason. Fix this by removing the gui package completely, then take out the
rootscripts from CMake and delete the CMakeLists there.