buildSystem: fix bug #24
fix problems with gcc 4.3:

* compiler flags that disable warnings are now checked, whether they
exist.

* std::complex accessors for real and imag part are kind of a
mess. gcc provides real() and imag() accessors that return references,
so they can be used in assigments. this, however, is an extension of
the C++ standard and not supported by Clang. with C++11 setters where
introduced, which, however, are not yet implemented in gcc 4.3. the
current solution is to check, whether the compiler provides a
gcc-compatible extension and if not use the standard C++11 setters.