Remove PWA2000.
Remove the pwa2000 subdirectory and clean all the build files. This involves
removing all references to pwa2000 from all the CMake files. Also remove all
the parts where the preprocessor variable USE_PWA2000 was involved. Compilation
is broken in this revision! The repercussions of this removal will be cleaned
up in the following commits.