Remove some old executables.
Remove several executables that could not be compiled anymore after the removal
of PWA2000. These seem not to have been too useful in any case.

Also remove the PWA2000 header includes from TFitBin and fitResult, obviously
they are not needed there.

This revision still does not build, i.e. the cleanup after the removal of
PWA2000 is not yet completed.