Final fixes to get rootpwa to build.
The most important fix is the transition in the TPWALikelihood from PWA2000's
integral class to the ampIntegralMatrix. It should be carefully reviewed if
this change is indeed done correctly.

Apart from that, delete the keygen directory completely, remove the
partialWeightWave in the generators and remove the TPWALikelihoodC along with
the executables which depend on it. Of course, also adjust all the CMakeLists
accordingly.