generator: fix a possible infinite loop
An infinite loop might happen at the comparison between 't' and 't0' (now at
line 187; should never happen, as 'tPrime' should always be negative if the
picker is working correctly). In any case pick new values for the mass and t'
if such a condition should occur.

The kinematics of X will now stay the some across attempts of the phase-space
decay, previously a new phi angle was obtained for each attempt.