Add parameterSpace class.
Add the parameterSpace class which can be used to perform different tasks
within the parameter space of the PWA fits, e.g. generate samples of amplitudes
which are uniformly distributed in parameter space and the like. So far, only
the calculation of the infinitesimal line and area elements is implemented and
this has not been tested yet. The next step will be implementing the uniform
and non uniform generation of amplitude samples with correct weighting.