Get parameterSpace class working.
Almost completely rewrite the parameterSpace class while hunting for bugs and
implementing performance improvements (mainly caching of calculated
quantities). Also add a test directory to src/ with one test testing the
parameterSpace. The parameterSpace class itself seems to be passing the most
rudimentary test: it will correctly calculate the surface area element of a
sphere when given the identity as integral matrix. Also, running it with a
integral matrix obtained from data, the results seem more or less reasonable.
However, a quantitative test remains to be found...