Improve performance of parameterSpace.
Improve the runtime of testParameterSpace by a factor of five by copying the
ampIntegralMatrix in the constructor, bypassing quite some checks which are
done in ampIntegralMatrix::elemente(). Also adjust one of the boundaries of one
of the loops so iterations which contribute zero are no longer executed.

In addition, remove commented code and clean up a little bit.