pyInterface: Improve extraction of std::vectors.
Improve the extraction of std::vector<>s from bp::objects by introducing a
templated function which does this in an error safe way. Change all instances
where this is needed to the new function. In the course of that, improve the
error reporting by making the C++ code raise python exceptions instead of the
usual "throw;", which should make error-handling more python-y when using the
bindings.