pyInterface: Add genpw.py.
Add a new script genpw.py. It is a python incarnation of the genpw executable.
It delivers the same results as its C++ counterpart and has the same
performance. However, there is a difference in the output when writing comgeant
files, where the precision of the C++ version for the vertex position is higher
for all but the first events. Presumably, this comes from some string stream
handling in generator::convertEventToComgeant which has to be looked at.