generator: Fix ASCII output of the generator.
Fix the handling of the precision in the functions handling the conversion of
events to ASCII and comgeant format by using the maxPrecision function of the
reportingUtils. Note that this slightly changes the output for the comgeant
format (the vertex positions are printed with high precision now for all
events).