Fix compilation after previous commit.
After the rename of the TCMatrix files to complexMatrix, fix the compilation by
changing all the instances in the code were TCMatrix was used. Also move the
complexMatrix to the rpwa namespace.

There was a problem which appeared in the stream operator of TComplex, which
stopped working because there was an operator already defined in std which
could handle TComplex and the compiler got mixed up. Fix this by removing the
stream operator for TComplex from reportingUtilsRoot.hpp. This changes the way
TComplex is printed from (Re, Im) to (Re,Imj), i.e. now TComplex is printed the
same as std::complex<>.