src: Rename TPWALikelihood to pwaLikelihood.
Rename the TPWALikelihood files to pwaLikelihood to avoid mix ups with root's
classes. This is the pure source and header file rename, no code was changed in
this commit to have a clean rename. This means compilation is broken in this
commit and will be fixed in the next one.