Adopt ROOT I/O to renaming of TCMatrix to complexMatrix
Increase the version number in the ClassDef of fitResult by one, the members of
this class have been changed when 'TCMatrix' was renamed to 'complexMatrix',
causing lengthy warnings from ROOT complaining about this.

In the dictionary ask that objects read from file as class 'TCMatrix' are to be
handled as 'complexMatrix'. As the internal structure of this class did not
change, but only its name, this should work.

As far as I could test, with this changes it is possible to read fitResults
created before the renaming with a current version of ROOTPWA.