Wave-specific mass-range for mass-dependent fit
It is now possible to set the fitted mass-range individually for each wave in
the mass-dependent fit. This required a number of changes to basically all
parts of the mass-dependent fitter.

In the configuration file one now has to specify the waves to be used in the
fit. At this place the mass-range can be specified with the options 'massLower'
and 'massUpper'. The example '5piMassDep.conf' has been adjusted accordingly
without defining any mass-range.

The default mass-range is now extracted from the mass-range covered by the bins
of the fit result, and is no longer hard-coded from 0 to 5000 MeV/c^2. Some
additional output has been added, so the result and correctness of this can
easily be checked.

The likelihood/chi^2 calculation has been adjusted to no take into account
mass-bins that are not in the range of each pair of waves, the highest lower
bound and the lowest higher bound define which mass-bins are used. A new
command-line option "-R" has been introduced controlling whether to plot the
result only in the specified mass-range or the total.

There is one issue if the plotting is not done of the complete mass-range. So
far the graphs are always filled at the index of the mass-bin, so that entries
with x- and y-values of 0 might exist in the beginning and end of the graph.
Combined with the drawing option this might create some strange loops in the
plots. Those are purely artifacts of the plotting, not of the fit itself.