Changed application of final-state mass-dependence
If the spin-density matrix is written as
                                 rho = T x T*
then the final-stat mass-dependence so far has been applied directly to the
product
                              rho = T x T* x fsmd
This has now been changed to
                         rho = (T x fsmd) x (T* x fsmd)
so the final-state mass-dependence now works on the "production amplitudes"
directly, and not on the spin-density matrix. This should make it simpler to in
the future experiment with directly fitting the production amplitudes, and not
the spin-density matrix.

(cherry picked from commit b08f825f4f5f4dd3cd6c296150a4173f6b053164)