pyInterface: Add helper functions to fitResult.
Add some helper functions to the bindings for the fitResult which are needed so
it can read from a TTree. This is very workaround-y but the best that is
available at the moment. It should be improved upon once it is clear how the
boost::python <-> PyROOT interaction actually works.