pyInterface: First version of genPseudoData.
Add all the code necessary to make genPseudoData do what it is supposed to do.
The script is now generating data, calculating the decay amplitudes on the fly,
and producing an "data-like" file (i.e. same data structure complete with the
weight) which can be used with hit and miss Monte Carlo to produce "artificial
data". It has been tested against calcAmplitudes.py/evtweight and produces the
same weights.

In this version, a lot of debug output is printed. This will be removed
in the next commit.