Cleaning command-line options
Moved the specification of the fit result from the mass-independent fit and the
specification of the files used for the plotting of the systematic errors from
the command-line to the configuration file. Removed the command-line switch
'-i', the name of the configuration file is now taken from the single
not-processed option.