Re-structure massDepFit.cc
Added a new class 'massDepFit' which is suppossed to do everything that is
currently done in the 'main' function. So far most of what the class does is
not yet used, so for the moment the same things are done twice. This will of
course be removed later on.

(cherry picked from commit aba4f52483398558c79a723d31da49cc2cfb2a53)