Change source of input for 'massDepFitLikeli'
The spin-density matrices used as an input for 'massDepFitLikeli' are now
created in the 'massDepFit' class. Cleaned the source of the 'massDepFitLikeli'
class, all calculations during the initialisation are now done outside, the
'init' method only sets some member variables.

The 'overlap' method of 'pwacompset' can now take an additional argument, the
bin number of the current mass bin. This might be used in the future to use the
value of the phase-space from an array, and not to have to calculate it from a
TGraph.