Plotting of phase-space integrals fixed
In one of the previous commits, the plotting of the phase-space integrals was
broken, as the memory layout of the integrals has changed when going from
vectors of vectors to a multi_array. For the calculation this has always been
done correctly.

(cherry picked from commit ffec7d6b30ca9381b03f5dbfe7a72ce8fd1585df)