Restructure 'massDepFitModel'
Started to restructure the 'massDepFitModel' class. Mainly renaming of methods
and variables, and some reindention.

The most notably change is adding a new method 'productionAmplitude', that
calculates the production amplitude for a given wave, and is used by the
'phase', 'intensity' and 'spinDensityMatrix' methods, instead of having an
independent implementation in each method.

For fixed final-state mass dependences pre-calculated values at the center of
the mass bins are stored and retrieved in case the 'idxMass' parameter has a
sensible value. This is the case during the fit itself.