Use class 'massDepFitComponent'
Use the class 'massDepFitComponent' as a base class for all components of the
mass-dependent fit. For this the class 'pwabkg' no longer inherits from
'pwacomponent' but only from 'massDepFitComponent', also 'pwacomponent'
inherits from 'massDepFitComponent'. 'massDepFitComponent' is a purely virtual
class providing an interface to the various channels. Parameter managment is to
be done by the inheriting classes.