Combine 'resonances' and 'background' in config file
Combine the two lists 'resonances' and 'background' into one big list. For this
it was neccessary to introduce a new field 'type', that can be either
'relativisticBreitWigner' or 'exponentialBackground', if that field is not set,
it is forced to 'relativisticBreitWigner'. The combining also required to put
couplings of the background into a 'decaychannel' group. For the moment it is
not checked, that a background component is only fitted to one wave.