Removed reading of fit model from main routine
Moved the reading of the fit model from the main routine to the 'massDepFit'
class which in turn calls the 'init' method of 'massDepFitCompontent' or
inheriting classes. Removed the 'jpc' field in the configuration file, it was
not used anywhere. The 'fix' field no has to be a boolean.

(cherry picked from commit e39638d72ca5b44a081b5350cb5fc68cb019a300)