First step in harmonizing releasePars with main
Started to harmonize what is done in releasePars to what is done in main, to
later be able to remove the code in main. Adjusted step sizes. In contrast to
anything written in comments and during running, in the first fit attempt, the
masses were not fixed.

(cherry picked from commit 48ea56655f0947d5197207e536e86dd11530195f)