Pass phase-space array index to pwachannel
As already done for the final-state mass-dependence in the 'massDepFitModel'
class, also pass an array index for the phase-space integrals used in the
'pwachannel' class. This array should be set to the index of the mass bin, if
the method should be evaluated at the mass bin center, and will speed up a lot.