Modified setting of parameters for fit
Modified the setting of parameters of the fit. This includes the changes to the
way the limits are now stored in the components, but mainly this is to allow
setting of variables that are unlimited, limited on only on side (upper or
lower limit, or limited on both sides.

(cherry picked from commit 8dc328b2c047b4cd67b3ba178ecdb7715514880a)