Methods to set/get parameters and couplings return number of set parameters
The methods to set or get resonance parameters and couplings in the 'component'
class now return the number of parameters set. So in 'model', the array of
parameters can now be assigned without an additional call to get this number.

(cherry picked from commit 4d61576c31dff66569115f82e4c8daf24d85e636)