massDepFit: enable reading of componenents for multiple bins
Started the third step to enable multi t'-bin fits, now the components for
multiple bins are read. However, a number of functions are still protected
by asserts to only run with a single bin, or have now a hard-coded bin 0.

To achieve reading the components, and later being able to do computations with
them, each method to get information from the fit model, like the calculation
of the production amplitude (and with this also to calculate the spin-density
matrix, phases, intensities) require an additional parameter for the bin for
which to do the calculation.

(cherry picked from commit e015291d4a23b4bedce0a687764bf9e0f13a24e1)