massDepFit: enable plotting for multiple bins
Started the fourth step to enable multi t'-bin fits, now the plotting can be
done for multiple bins. If multiple bins are present in the fit, then each bin
it put into its own directory in the ROOT file created. The information from
the relevant bin are used for all plotted variables.