massDepFit: fitting in multiple bins
Finished the steps to perform a multi t'-bin fit. Finally the 'likelihood'
class has been adopted to be able to digest an additional dimension in the
spin-density matrices. Inside the evaluation of the likelihood, a loop over all
bins is done. While initializing the parameters for the minimizer now include
the bin number into the parameter name.