massDepFit: control free parameters via configuration file
Allow to define which parameters to free in the minimizer at which minimizer
step via the configuration file. A new option 'freeparameters' was added, which
is an array of strings. If the option is not specified, the old behaviour is
done. Each entry in the array defined a space or comma separated list of
free parameters, the options are '', '*', component names, parameter names and
complete variable names (component name + "__" + parameter name). The couplings
are always free.

(cherry picked from commit 5c95cdc0e6d8b15b14cc595206779d663de9e55c)