massDepFit: calculate a chi2 value in plotting-only mode
Also calculate a chi2 value from the parameters in the configuration file in
the plotting-only mode. This might be helpful, if one forgot to note the value
during the fit and the fit was not valid (then the chi2 is not stored in the
configuration file).

(cherry picked from commit 9d4636e04b5f761478a92badc8624a5d5b00f34a)