massDepFit: production amplitudes in massDepFitLikeli
Make the production amplitudes and the covariances available in the
'massDepFitLikeli' class, but do not use them yet.

Create a new method 'DoEvalSpinDensityMatrix' evaluating the chi2 for the fit
to the spin-density matrix, moved the code from the old 'DoEval' method there.
'DoEval' for the moment only passes the fit parameters to the model and calls
the new method.

(cherry picked from commit 2337a81a16e884ddf910d288cb4feb15f23449d2)