massDepFit: bool for real anchor wave is kept only locally
Started to change the code to be able to also handle non-real anchor waves. The
idea is to rotate the production amplitudes (and the covariance matrix) such
that the anchor wave is real again. This means neither the method to get the
number of parameters, nor the method to evaluate the chi2 need to know whether
the anchor wave was real or not, but can rely on it being real.

Also added a test that if the anchor wave is not real, no other wave is real.

(cherry picked from commit c7493bce6be035ed4c84257c0c8f1f85fd53f318)