massDepFit: test that anchor wave is non-zero
Make sure that the anchor wave is non-zero over the fitted mass range (i.e. due
to a mass threshold in the mass-independent fit).

Also limit the mass-range in which the production amplitude and covariance
matrix are rotated to the fit range.

(cherry picked from commit 41adf285891486491ebeee586ed41a0d3d03924b)