massDepFit: skip zero waves
During the initialization of the fit to the production amplitudes skip waves
that are zero (amplitude and covariance, i.e. due to mass threshold). Those
waves are skipped as otherwise the covariance matrix cannot be inverted. For
the final matrix, zero columns and rows are inserted for the skipped waves.

(cherry picked from commit 8f5e201b81576495ac55b068015311fb8e9dde40)