massDepFit: simplify Bowler parameterization
The phase-space is already stored in the channel, so why store it also in this
class? In principle this should also work for multiple channels, however, for
the moment this is not allowed, probably one should then introduce branching
ratios (as it is done for the relativistic Breit-Wigner).

(cherry picked from commit 54cfe811419e315b0cf73a7558573f4414914622)