massDepFit: variables of final-state mass-dependence need to be float
Previously it was only checked that the variable are a number
(libconfig::Setting::isNumber), this means either float or integer. This might
cause problems if a value from the fit (typically a float) needs to be written
back, but the variable previously was an integer.

(cherry picked from commit bd8718df285ad9650b35c89d04b953779e66cecf)