massDepFit: count number of couplings
In the 'init' method of the 'component' class, count the number of couplings
read from the configuration file. The number of parameters then has not to be
guessed from the number of channels, but can directly be calculated from the
new number.

(cherry picked from commit 3113c45c58236c9d88912d75d99b2af31ccdcae4)