massDepFit: conditional reading of couplings
While reading the components from the configuration file, the read reading of
the couplings is controlled via a boolean flag, which - for the moment - is
always true. If no couplings are to be read, the default value 1 is used.

(cherry picked from commit c6f2062f6d05a81d3ada7e3e08c7fb1533358c7b)