massDepFit: introduce mapping of channel to couplings
Introduce a mapping of the decay channel of a component to the couplings.
Accordingly the updating of the parameters, and the addition of the fit
parameters if conditioned by the equality of the index of the decay channel
and the index of the couplings.
the index of the decay channel agrees with the com