massDepFit: enable branchings in the fit
Finally enable the using of the branchings in the fit. Take into account the
number of branchings for the number of parameters in the model. While setting
up the fitparameters, create the name of the couplings only from the quantum
number part of the wave name. The first branching of each component is fixed.