massDepFit: move reading of config from main to 'massDepFit'
Move the reading of the configuration file, and the initialization of the
fit-model and the likelihood, from the main routinge to new 'readConfig' and
'init' methods of the 'massDepFit' class. Removed a number of methods
previously needed to access internal data of the 'massDepFit' class.

(cherry picked from commit 81a2370c6e5eca3ebed6a150334f120ed303e038)