evtweight: change values of reflectivity
The reflectivity for a wave (except the flat wave) is now either -1 or +1,
while previously it has been 1 or 0, respectively. It seems that in ancient
times, the value of the reflectivity was used as on offset in an array, as this
is no longer the case, the real value can now be used.