evtweight: weighting for non rank-1 fits
The calculation of weights should now also work for non rank-1 fits.

While reading the input from the 'fitResult' object do not loop over the
waves, but over the production amplitudes instead. The wave name (still
required for, e.g., the loading of the amplitudes) is also extracted
from the 'fitResult' object, and index for the corresponding wave is
kept for each production amplitude.

Also loop over the production amplitudes in the calculation.

For the moment individual weights are only stores for production
amplitudes, not yet waves.