evtweight: add weight for waves
Add the weight per wave as the sum of the weights of the production
amplitudes. The weight for the production amplitudes is only added to
the tree, if the rank is higher than one. In that case, the weight per
wave is in any case the weight per production amplitude.