evtweight: enable reading of ROOT amplitudes
Enabled the reading of amplitudes from ROOT files/trees. This can be
switched on with the command line option '-R'. The implementation of
reading the files is very much like the one in the 'pwaLikelihood' and
'ampIntegralMatrix' classes. As for the fitter, all amplitudes have to
be of the same format.