evtweight: remove event input files
The event files are no longer required. The numer of events to loop over
is now extracted from the number of values in the amplitude files. Along
with this change, the events are also no longer copied to the output
file, instead a tree (by default) called 'rootPwaWeightTree' is created
and filled with only the weights. This tree should the be used as a
friend of the event tree.