utilities: do not print string at NULL
The 'getName' method of a 'Setting' object of Libconfig might actually
return NULL, e.g. for an unamed group (which might happen if one wants
to store several similar settings in a list of groups). If such a NULL
is encountered do not attempt to print it, but print '???' instead.