pyInterface: use friend tree in 'plotAngles.py'
Make 'plotAngles.py' compatible to the recent changes to 'evtweight'.
'evtweight''s output does no longer contain a copy of the events, so
the weights have to be added as a friend tree to the event tree.

In principle adding a friend that still contains the weight and the
events should work. In that case the weight and events should be read
from the two different trees.