src: Store acceptance integrals in fit result.
In order to improve the evidence calculation and get rid of the approximation
for the volume of the parameter space, store the normalization integral with
acceptance in the fit result file. This way, it can be accessed later when the
parameter space is calculated, which is the part which will be added next.

The version number in the ClassDef had to be raised for the fitResult class.