pyInterface: Change waves logic in genPseudoData.
To determine which waves to use when generating events, do not use the keyfiles
anymore but read the wave names from the fit result file. This makes much more
sense because the keyfile pool can remain the same and may be much bigger than
the actual waves which are given in the fit result file.