pyInterface: Fix getBestFitResult for multiple entries.
The 'getBestFitResult' functioned contained an error that occured for
trees with more than on entry, in this case the variable 'bestLike' was
not defined. In total there were three variables 'bestLike',
'bestLikeli', and 'logLikeBest' where probably there should have been
only one.

Corrected this and also renamed 'massBinCenterBest' to 'bestMass', and
'logLike' to 'likeli' to improve the resemblance between the variables
for the current entry and the best.