amplitude: Fix bug in phaseSpaceIntegral.
Fix a subtle bug in the phaseSpaceIntegral which could occur due to the pointer
bookkeeping changing between different calculation runs (occurring when using
the calcAmplitudes.py) with the singleton not being reset.

Note that the whole phaseSpaceIntegral business is quite "hacky" at the moment
and should be properly re-designed at some point.