utilities: Move functions printing environment.
Move the three functions printing information about the environment to their
own files, which are now compiled into the utilities library, which did not
exist before. This is done as a preparation to fix the problem that the git
hash was not updated unless all the cmake files were deleted.