buildSystem: Use correct git hash.
Fix bug #7 [bugs:#7] with some cmake magic. This should always use the most
recent value for the git hash, the user name and the system host name without
recompiling if these values did not change. As far as I can tell, the other
variables which are defined will still be cached and might be out-of-date, a
problem which is left for a future update.