utilities: Add metadata object for data.
Add a new file to the utilities directory which contains the metadata object to
be used in the data files. The same file should also contain the converter
class which handles the filling of the metadata, but this is not implemented
yet. Add a utilities test which does nothing except including the newly created
files so that it is compiled somewhere.