utilities: Add executables for meta data.
Add two executables for the handling of meta data. The first one takes an
rootpwa file in the format that was used up to now and produces a new file in
the new format, i.e. including the metadata. The second print meta data
information for a given file.

Also, alter the structure a little bit: the dataMetadata class gets its own
source and header file together with a dictionary for that while the
rootpwaDataFileWriter stay in dataConverterUtils.hpp. This is not very nice,
but will be leaned up in the next commit(s).