storageFormats: Revise eventStorage.
Completely revise the eventStorage: there is now an eventStorage class which is
the only thing stored in the event files. With this implementation, hadd is
functional and the hadded files have valid hashes and concatenated users
strings. The eventFileWriter and the printMetadata executable have been adapted
to these new additions.

Unfortunately, the event tree is not saved as a pointer in the eventStorage
object, as there seem to be problems with saving the tree to disk in that case.
The impact of this as well as possible alternative solutions will be
investigated.