ROOT6 compatibility during build process
With ROOT6 'rootcint'/'rootcling' does not define __CINT__ for some (all?)
parsing of the header files, so some additional include directories have
to be specified at this place.

Also 'private' and 'protected' are no longer defined as 'public' during
the compilation of the dictionaries, so all methods called for scheme
evolution rules really have to be public. In case of the 'waveDescription'
class the previously used private method has been replaced by a chain of
two already public methods.