buildSystem: changes to dictionary creation
ROOT6 requires some changes to the way the dictionaries are created.
'rootcling' (what was formerly called 'rootcint' does not only create the
dictionary (only the source is created for ROOT6, the header is no longer
necessary), but also creates some pre-compiled header stuff, and the
rootmap. This additional stuff needs to end up in the same place as the
dynamic library. Also for ROOT5 a rootmap is now created.

This required some changes to the arguments of the Cmake macro creating
the dictionary.