storageFormats: Delete storage class again.
Remove the storage class again in favor of the solution with a TTree in the
event file together with a metadata object. As this does not work with hadd,
provide a new executable mergeDatafiles which does as advertised and merges
different data files into one, thereby correctly setting the hash and the other
variables in the matadata.