pyInterface: Fix bug in genPseudoData.py.
Fix a bug in genPseudoData: the location of the keyfile directory was
hard-coded while the keyfile command line argument did not have any effect.
Change the command line argument so that now the keyfile directory has to be
passed instead of the single keyfiles. As the fitResult which is given
determines the waves which are used, the script will now take the keyfiles it
needs form this new argument.

Also harmonize the capitalization when printing and remove module names from
printing function, as they are renamed at the beginning of the script.

(cherry picked from commit 510d28f38e8bd25d0825da89858a324bd001ee87)