UserSpace: clean 3pi.-00 and pi_2eta.-00 directories
Remove the prefixes '3pin' and 'pi-2eta' from the filenames, this is
already included in the directory name. Adjust the path to the particle
data table in GenerateKeyfiles.sh, compared to the old 'keyfiles'
directory, there is one more intermediate directory. Added an example
generator configuration file also for pi-etaeta.