UserSpace: modernize 3pi.--+ directory
Modernize the directory for the pi- pi- pi+ channel. Updated the example
generator configuration to work with the current generator, also updated
the target description to a hydrogen target. Removed the existing
keyfiles and wavelist, updated the template for the wave set generator
and added a number of wavelist of previous experiments. Also added a
script calling the wave set generator and taking care of creating the
special mass dependencies.

Basically this is a copy of the directory for pi- pi0 pi0 with changed
pion charges.

(cherry picked from commit 9450af8ef8bda6d3aaa3fe70c3b71051ede34d92)