generators: Fix issues with plotting weighted events for pi- pi0 pi0
Use the correct index for the particle name, when printing the error
message that a particle from the file is not found in the particle data
table.

Reduce scope of the variable used to keep track of the maximal and
average weight, the average weight is now calculated correctly and no
longer off by 1/nrEvents.