Add high level interface.
Add the high level interface directory which should house high level rootpwa
functions which are then exported to python and should form the foundation of
the user interface scripts. At the moment, add a calcAmplitude function which
is mostly copied from evtTreeHelper's processTree but uses the new
eventMetadata object. It has only been tested that it does something, but the
comparison between the old amplitude calculation executables and scripts is not
done yet.